refactor: simple feature structure
