+++ lib/rearchitecture/features/pillar_rewards_history/pillar_rewards_history.dart
@@ -1,2 +1,2 @@
 export './bloc/pillar_rewards_history_bloc.dart';
-export 'view/view.dart';
+export './view/pillar_rewards_card.dart';
