--- lib/rearchitecture/features/pillar_rewards_history/view/view.dart
@@ -1,2 +1,0 @@
-export './widgets/widgets.dart';
-export 'pillar_rewards_card.dart';
