--- lib/rearchitecture/features/pillar_rewards_history/view/widgets/widgets.dart
@@ -1,1 +1,0 @@
-export 'pillar_rewards_chart.dart';
