+++ lib/rearchitecture/features/sentinel_rewards_history/sentinel_rewards_history.dart
@@ -1,2 +1,2 @@
 export './bloc/sentinel_rewards_history_bloc.dart';
-export 'view/view.dart';
+export './view/sentinel_rewards_card.dart';
