--- lib/rearchitecture/features/sentinel_rewards_history/view/view.dart
@@ -1,2 +1,0 @@
-export './widgets/widgets.dart';
-export 'sentinel_rewards_card.dart';
