--- lib/rearchitecture/features/sentinel_rewards_history/view/widgets/widgets.dart
@@ -1,1 +1,0 @@
-export 'sentinel_rewards_chart.dart';
